test: parallelize benchmark cases across worker processes
Split the benchmark harness into an orchestrator
(BenchmarkTestCases.benchmark.mjs) and a jest-worker worker
(test/harness/benchmark/benchmark.worker.mjs). Time and simulation runs
now dispatch each benchmark case to a pool of worker processes instead
of running every case sequentially, cutting wall time on the sharded CI
jobs. Memory mode stays single-threaded in the main process so libuv and
heap accounting remain deterministic.

Move the shared create-tree helper under test/harness/benchmark and keep
the harness linted.
