+++ package.json
@@ -160,6 +160,7 @@
     "jest-environment-node": "^30.3.0",
     "jest-junit": "^17.0.0",
     "jest-snapshot": "^30.3.0",
+    "jest-worker": "^30.2.0",
     "json-loader": "^0.5.7",
     "json5": "^2.1.3",
     "less": "^4.6.4",
