+++ test/benchmarkCases/many-modules-commonjs/options.mjs
@@ -1,7 +1,7 @@
 import fs from "fs/promises";
 import { dirname, resolve } from "path";
 import { fileURLToPath } from "url";
-import createTree from "../_helpers/create-tree.mjs";
+import createTree from "../../harness/benchmark/create-tree.mjs";
 
 export async function setup() {
 	const __dirname = dirname(fileURLToPath(import.meta.url));
